requirements.txt
rich==13.8.1
revolt.py==0.2.0
python-dotenv==1.0.1
aiohttp-client-cache[aiosqlite]==0.11.1
aiosqlite==0.20.0